+++ packages/atomic-react/src/components/commerce/components.ts
@@ -58,331 +58,339 @@ import {
 import {createComponent} from '@lit/react';
 import React from 'react';
 
-export const AtomicAriaLive = createComponent({
+export const AtomicAriaLive = /*@__PURE__*/ createComponent({
   tagName: 'atomic-aria-live',
   react: React,
   elementClass: LitAtomicAriaLive,
 });
 
-export const AtomicCommerceBreadbox = createComponent({
+export const AtomicCommerceBreadbox = /*@__PURE__*/ createComponent({
   tagName: 'atomic-commerce-breadbox',
   react: React,
   elementClass: LitAtomicCommerceBreadbox,
 });
 
-export const AtomicCommerceCategoryFacet = createComponent({
+export const AtomicCommerceCategoryFacet = /*@__PURE__*/ createComponent({
   tagName: 'atomic-commerce-category-facet',
   react: React,
   elementClass: LitAtomicCommerceCategoryFacet,
 });
 
-export const AtomicCommerceDidYouMean = createComponent({
+export const AtomicCommerceDidYouMean = /*@__PURE__*/ createComponent({
   tagName: 'atomic-commerce-did-you-mean',
   react: React,
   elementClass: LitAtomicCommerceDidYouMean,
 });
 
-export const AtomicCommerceFacet = createComponent({
+export const AtomicCommerceFacet = /*@__PURE__*/ createComponent({
   tagName: 'atomic-commerce-facet',
   react: React,
   elementClass: LitAtomicCommerceFacet,
 });
 
-export const AtomicCommerceFacets = createComponent({
+export const AtomicCommerceFacets = /*@__PURE__*/ createComponent({
   tagName: 'atomic-commerce-facets',
   react: React,
   elementClass: LitAtomicCommerceFacets,
 });
 
-export const AtomicCommerceInterface = createComponent({
+export const AtomicCommerceInterface = /*@__PURE__*/ createComponent({
   tagName: 'atomic-commerce-interface',
   react: React,
   elementClass: LitAtomicCommerceInterface,
 });
 
-export const AtomicCommerceLayout = createComponent({
+export const AtomicCommerceLayout = /*@__PURE__*/ createComponent({
   tagName: 'atomic-commerce-layout',
   react: React,
   elementClass: LitAtomicCommerceLayout,
 });
 
-export const AtomicCommerceLoadMoreProducts = createComponent({
+export const AtomicCommerceLoadMoreProducts = /*@__PURE__*/ createComponent({
   tagName: 'atomic-commerce-load-more-products',
   react: React,
   elementClass: LitAtomicCommerceLoadMoreProducts,
 });
 
-export const AtomicCommerceNoProducts = createComponent({
+export const AtomicCommerceNoProducts = /*@__PURE__*/ createComponent({
   tagName: 'atomic-commerce-no-products',
   react: React,
   elementClass: LitAtomicCommerceNoProducts,
 });
 
-export const AtomicCommerceNumericFacet = createComponent({
+export const AtomicCommerceNumericFacet = /*@__PURE__*/ createComponent({
   tagName: 'atomic-commerce-numeric-facet',
   react: React,
   elementClass: LitAtomicCommerceNumericFacet,
 });
 
-export const AtomicCommercePager = createComponent({
+export const AtomicCommercePager = /*@__PURE__*/ createComponent({
   tagName: 'atomic-commerce-pager',
   react: React,
   elementClass: LitAtomicCommercePager,
 });
 
-export const AtomicCommerceProductList = createComponent({
+export const AtomicCommerceProductList = /*@__PURE__*/ createComponent({
   tagName: 'atomic-commerce-product-list',
   react: React,
   elementClass: LitAtomicCommerceProductList,
 });
 
-export const AtomicCommerceProductsPerPage = createComponent({
+export const AtomicCommerceProductsPerPage = /*@__PURE__*/ createComponent({
   tagName: 'atomic-commerce-products-per-page',
   react: React,
   elementClass: LitAtomicCommerceProductsPerPage,
 });
 
-export const AtomicCommerceQueryError = createComponent({
+export const AtomicCommerceQueryError = /*@__PURE__*/ createComponent({
   tagName: 'atomic-commerce-query-error',
   react: React,
   elementClass: LitAtomicCommerceQueryError,
 });
 
-export const AtomicCommerceQuerySummary = createComponent({
+export const AtomicCommerceQuerySummary = /*@__PURE__*/ createComponent({
   tagName: 'atomic-commerce-query-summary',
   react: React,
   elementClass: LitAtomicCommerceQuerySummary,
 });
 
-export const AtomicCommerceRecommendationInterface = createComponent({
-  tagName: 'atomic-commerce-recommendation-interface',
-  react: React,
-  elementClass: LitAtomicCommerceRecommendationInterface,
-});
+export const AtomicCommerceRecommendationInterface =
+  /*@__PURE__*/ createComponent({
+    tagName: 'atomic-commerce-recommendation-interface',
+    react: React,
+    elementClass: LitAtomicCommerceRecommendationInterface,
+  });
 
-export const AtomicCommerceRecommendationList = createComponent({
+export const AtomicCommerceRecommendationList = /*@__PURE__*/ createComponent({
   tagName: 'atomic-commerce-recommendation-list',
   react: React,
   elementClass: LitAtomicCommerceRecommendationList,
 });
 
-export const AtomicCommerceRefineModal = createComponent({
+export const AtomicCommerceRefineModal = /*@__PURE__*/ createComponent({
   tagName: 'atomic-commerce-refine-modal',
   react: React,
   elementClass: LitAtomicCommerceRefineModal,
 });
 
-export const AtomicCommerceRefineToggle = createComponent({
+export const AtomicCommerceRefineToggle = /*@__PURE__*/ createComponent({
   tagName: 'atomic-commerce-refine-toggle',
   react: React,
   elementClass: LitAtomicCommerceRefineToggle,
 });
 
-export const AtomicCommerceSearchBox = createComponent({
+export const AtomicCommerceSearchBox = /*@__PURE__*/ createComponent({
   tagName: 'atomic-commerce-search-box',
   react: React,
   elementClass: LitAtomicCommerceSearchBox,
 });
 
-export const AtomicCommerceSearchBoxInstantProducts = createComponent({
-  tagName: 'atomic-commerce-search-box-instant-products',
-  react: React,
-  elementClass: LitAtomicCommerceSearchBoxInstantProducts,
-});
+export const AtomicCommerceSearchBoxInstantProducts =
+  /*@__PURE__*/ createComponent({
+    tagName: 'atomic-commerce-search-box-instant-products',
+    react: React,
+    elementClass: LitAtomicCommerceSearchBoxInstantProducts,
+  });
 
-export const AtomicCommerceSearchBoxQuerySuggestions = createComponent({
-  tagName: 'atomic-commerce-search-box-query-suggestions',
-  react: React,
-  elementClass: LitAtomicCommerceSearchBoxQuerySuggestions,
-});
+export const AtomicCommerceSearchBoxQuerySuggestions =
+  /*@__PURE__*/ createComponent({
+    tagName: 'atomic-commerce-search-box-query-suggestions',
+    react: React,
+    elementClass: LitAtomicCommerceSearchBoxQuerySuggestions,
+  });
 
-export const AtomicCommerceSearchBoxRecentQueries = createComponent({
-  tagName: 'atomic-commerce-search-box-recent-queries',
-  react: React,
-  elementClass: LitAtomicCommerceSearchBoxRecentQueries,
-});
+export const AtomicCommerceSearchBoxRecentQueries =
+  /*@__PURE__*/ createComponent({
+    tagName: 'atomic-commerce-search-box-recent-queries',
+    react: React,
+    elementClass: LitAtomicCommerceSearchBoxRecentQueries,
+  });
 
-export const AtomicCommerceSortDropdown = createComponent({
+export const AtomicCommerceSortDropdown = /*@__PURE__*/ createComponent({
   tagName: 'atomic-commerce-sort-dropdown',
   react: React,
   elementClass: LitAtomicCommerceSortDropdown,
 });
 
-export const AtomicCommerceText = createComponent({
+export const AtomicCommerceText = /*@__PURE__*/ createComponent({
   tagName: 'atomic-commerce-text',
   react: React,
   elementClass: LitAtomicCommerceText,
 });
 
-export const AtomicCommerceTimeframeFacet = createComponent({
+export const AtomicCommerceTimeframeFacet = /*@__PURE__*/ createComponent({
   tagName: 'atomic-commerce-timeframe-facet',
   react: React,
   elementClass: LitAtomicCommerceTimeframeFacet,
 });
 
-export const AtomicComponentError = createComponent({
+export const AtomicComponentError = /*@__PURE__*/ createComponent({
   tagName: 'atomic-component-error',
   react: React,
   elementClass: LitAtomicComponentError,
 });
 
-export const AtomicFocusTrap = createComponent({
+export const AtomicFocusTrap = /*@__PURE__*/ createComponent({
   tagName: 'atomic-focus-trap',
   react: React,
   elementClass: LitAtomicFocusTrap,
 });
 
-export const AtomicGeneratedAnswerFeedbackModal = createComponent({
-  tagName: 'atomic-generated-answer-feedback-modal',
-  react: React,
-  elementClass: LitAtomicGeneratedAnswerFeedbackModal,
-});
+export const AtomicGeneratedAnswerFeedbackModal = /*@__PURE__*/ createComponent(
+  {
+    tagName: 'atomic-generated-answer-feedback-modal',
+    react: React,
+    elementClass: LitAtomicGeneratedAnswerFeedbackModal,
+  }
+);
 
-export const AtomicIcon = createComponent({
+export const AtomicIcon = /*@__PURE__*/ createComponent({
   tagName: 'atomic-icon',
   react: React,
   elementClass: LitAtomicIcon,
 });
 
-export const AtomicLayoutSection = createComponent({
+export const AtomicLayoutSection = /*@__PURE__*/ createComponent({
   tagName: 'atomic-layout-section',
   react: React,
   elementClass: LitAtomicLayoutSection,
 });
 
-export const AtomicNumericRange = createComponent({
+export const AtomicNumericRange = /*@__PURE__*/ createComponent({
   tagName: 'atomic-numeric-range',
   react: React,
   elementClass: LitAtomicNumericRange,
 });
 
-export const AtomicProduct = createComponent({
+export const AtomicProduct = /*@__PURE__*/ createComponent({
   tagName: 'atomic-product',
   react: React,
   elementClass: LitAtomicProduct,
 });
 
-export const AtomicProductChildren = createComponent({
+export const AtomicProductChildren = /*@__PURE__*/ createComponent({
   tagName: 'atomic-product-children',
   react: React,
   elementClass: LitAtomicProductChildren,
 });
 
-export const AtomicProductDescription = createComponent({
+export const AtomicProductDescription = /*@__PURE__*/ createComponent({
   tagName: 'atomic-product-description',
   react: React,
   elementClass: LitAtomicProductDescription,
 });
 
-export const AtomicProductExcerpt = createComponent({
+export const AtomicProductExcerpt = /*@__PURE__*/ createComponent({
   tagName: 'atomic-product-excerpt',
   react: React,
   elementClass: LitAtomicProductExcerpt,
 });
 
-export const AtomicProductFieldCondition = createComponent({
+export const AtomicProductFieldCondition = /*@__PURE__*/ createComponent({
   tagName: 'atomic-product-field-condition',
   react: React,
   elementClass: LitAtomicProductFieldCondition,
 });
 
-export const AtomicProductImage = createComponent({
+export const AtomicProductImage = /*@__PURE__*/ createComponent({
   tagName: 'atomic-product-image',
   react: React,
   elementClass: LitAtomicProductImage,
 });
 
-export const AtomicProductLink = createComponent({
+export const AtomicProductLink = /*@__PURE__*/ createComponent({
   tagName: 'atomic-product-link',
   react: React,
   elementClass: LitAtomicProductLink,
 });
 
-export const AtomicProductMultiValueText = createComponent({
+export const AtomicProductMultiValueText = /*@__PURE__*/ createComponent({
   tagName: 'atomic-product-multi-value-text',
   react: React,
   elementClass: LitAtomicProductMultiValueText,
 });
 
-export const AtomicProductNumericFieldValue = createComponent({
+export const AtomicProductNumericFieldValue = /*@__PURE__*/ createComponent({
   tagName: 'atomic-product-numeric-field-value',
   react: React,
   elementClass: LitAtomicProductNumericFieldValue,
 });
 
-export const AtomicProductPrice = createComponent({
+export const AtomicProductPrice = /*@__PURE__*/ createComponent({
   tagName: 'atomic-product-price',
   react: React,
   elementClass: LitAtomicProductPrice,
 });
 
-export const AtomicProductRating = createComponent({
+export const AtomicProductRating = /*@__PURE__*/ createComponent({
   tagName: 'atomic-product-rating',
   react: React,
   elementClass: LitAtomicProductRating,
 });
 
-export const AtomicProductSectionActions = createComponent({
+export const AtomicProductSectionActions = /*@__PURE__*/ createComponent({
   tagName: 'atomic-product-section-actions',
   react: React,
   elementClass: LitAtomicProductSectionActions,
 });
 
-export const AtomicProductSectionBadges = createComponent({
+export const AtomicProductSectionBadges = /*@__PURE__*/ createComponent({
   tagName: 'atomic-product-section-badges',
   react: React,
   elementClass: LitAtomicProductSectionBadges,
 });
 
-export const AtomicProductSectionBottomMetadata = createComponent({
-  tagName: 'atomic-product-section-bottom-metadata',
-  react: React,
-  elementClass: LitAtomicProductSectionBottomMetadata,
-});
+export const AtomicProductSectionBottomMetadata = /*@__PURE__*/ createComponent(
+  {
+    tagName: 'atomic-product-section-bottom-metadata',
+    react: React,
+    elementClass: LitAtomicProductSectionBottomMetadata,
+  }
+);
 
-export const AtomicProductSectionChildren = createComponent({
+export const AtomicProductSectionChildren = /*@__PURE__*/ createComponent({
   tagName: 'atomic-product-section-children',
   react: React,
   elementClass: LitAtomicProductSectionChildren,
 });
 
-export const AtomicProductSectionDescription = createComponent({
+export const AtomicProductSectionDescription = /*@__PURE__*/ createComponent({
   tagName: 'atomic-product-section-description',
   react: React,
   elementClass: LitAtomicProductSectionDescription,
 });
 
-export const AtomicProductSectionEmphasized = createComponent({
+export const AtomicProductSectionEmphasized = /*@__PURE__*/ createComponent({
   tagName: 'atomic-product-section-emphasized',
   react: React,
   elementClass: LitAtomicProductSectionEmphasized,
 });
 
-export const AtomicProductSectionMetadata = createComponent({
+export const AtomicProductSectionMetadata = /*@__PURE__*/ createComponent({
   tagName: 'atomic-product-section-metadata',
   react: React,
   elementClass: LitAtomicProductSectionMetadata,
 });
 
-export const AtomicProductSectionName = createComponent({
+export const AtomicProductSectionName = /*@__PURE__*/ createComponent({
   tagName: 'atomic-product-section-name',
   react: React,
   elementClass: LitAtomicProductSectionName,
 });
 
-export const AtomicProductSectionVisual = createComponent({
+export const AtomicProductSectionVisual = /*@__PURE__*/ createComponent({
   tagName: 'atomic-product-section-visual',
   react: React,
   elementClass: LitAtomicProductSectionVisual,
 });
 
-export const AtomicProductText = createComponent({
+export const AtomicProductText = /*@__PURE__*/ createComponent({
   tagName: 'atomic-product-text',
   react: React,
   elementClass: LitAtomicProductText,
 });
 
-export const AtomicTimeframe = createComponent({
+export const AtomicTimeframe = /*@__PURE__*/ createComponent({
   tagName: 'atomic-timeframe',
   react: React,
   elementClass: LitAtomicTimeframe,